Update SiD_o2_v01.xml
Update to be used in conjunction with ECalBarrel driver update. Changed "ECalBarrel_rmin=1264" from 1265 so that barrel fits in the allotted space. Note that this value disagrees with the engineering design, but extends into empty space in front of the ECal (see discussion on https://silicondetector.slack.com/messages/detector_geometry/ )

This change would need to be applied to SiD_o2_v02 as well